* thread.c (rb_fd_rcopy): suppress warnings on mingw32.
* win32/win32.c (overlapped_socket_io, recvmsg, sendmsg, dupfd):
  ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@33621 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ thread.c
@@ -2471,7 +2471,7 @@ rb_fd_rcopy(fd_set *dst, rb_fdset_t *src)
 
     /* we assume src is the result of select() with dst, so dst should be
      * larger or equal than src. */
-    if (max > FD_SETSIZE || max > dst->fd_count) {
+    if (max > FD_SETSIZE || (UINT)max > dst->fd_count) {
 	rb_raise(rb_eArgError, "too large fdsets");
     }
